Refactor HistoryScreen
