+++ lib/widgets/transactions_list.dart
@@ -1,0 +1,45 @@
+import 'package:flutter/material.dart';
+import 'package:animations/animations.dart';
+import 'package:provider/provider.dart';
+
+import '../card_items/transaction_card.dart';
+import '../providers/transactions.dart';
+import '../screens/transaction_details_screen.dart';
+
+// Used in HistoryScreen.
+
+class TransactionsList extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    final transactions = Provider.of<Transactions>(context).items;
+
+    if (transactions.length == 0) {
+      return const Center(
+        child: Text('Start adding some transactions!'),
+      );
+    }
+
+    return ListView.builder(
+      itemCount: transactions.length,
+      itemBuilder: (_, index) {
+        return Column(
+          children: <Widget>[
+            OpenContainer(
+              closedShape: const BeveledRectangleBorder(),
+              closedElevation: 0,
+              closedBuilder: (_, __) {
+                return TransactionCard(transaction: transactions[index]);
+              },
+              openBuilder: (_, __) {
+                return TransactionDetailsScreen(
+                  transactionId: transactions[index].id,
+                );
+              },
+            ),
+            const Divider(height: 1),
+          ],
+        );
+      },
+    );
+  }
+}
